refactor: remove unused RunCmdNode type definitions and improve SaveTextNode component structure
feat: implement useNoteNode composable for managing note functionality

feat: create useOpenFileNode composable for handling file opening logic

feat: add usePythonRunner composable for executing Python scripts with dependency management

feat: introduce useRepoConcat composable for concatenating repository files

feat: develop useSaveTextNode composable for saving text to files with error handling
