+++ frontend/src/components/NoteNode.vue
@@ -21,12 +21,9 @@
       <button @click.prevent="decreaseFontSize">-</button>
       <button @click.prevent="increaseFontSize">+</button>
     </div>
-
     <div :style="data.labelStyle" class="node-label">{{ data.type }}</div>
-
     <Handle style="width:12px; height:12px" v-if="data.hasInputs" type="target" position="left" id="input" />
     <Handle style="width:12px; height:12px" v-if="data.hasOutputs" type="source" position="right" id="output" />
-
     <div
       class="text-container"
       ref="textContainer"
@@ -53,19 +50,11 @@
     />
   </div>
 </template>
-
 <script setup>
-import { ref, computed, watch, nextTick, onMounted } from 'vue'
-import { Handle, useVueFlow } from '@vue-flow/core'
+import { watch, onMounted } from 'vue'
+import { Handle } from '@vue-flow/core'
 import { NodeResizer } from '@vue-flow/node-resizer'
-
-async function run() {
-  return
-}
-
-onMounted(() => {
-  props.data.run = run
-})
+import { useNoteNode } from '@/composables/useNoteNode'
 
 const props = defineProps({
   id: {
@@ -106,113 +95,37 @@ const emit = defineEmits([
   'resize'
 ])
 
-const noteText = computed({
-  get: () => props.data.inputs.note,
-  set: (value) => {
-    props.data.inputs.note = value
-  }
+// Use the note node composable
+const {
+  noteText,
+  isHovered,
+  customStyle,
+  currentFontSize,
+  textContainer,
+  currentColor,
+  resizeHandleStyle,
+  increaseFontSize,
+  decreaseFontSize,
+  handleScroll,
+  onResize,
+  cycleColor,
+  run,
+} = useNoteNode(props, emit)
+
+// Set run function on component mount
+onMounted(() => {
+  props.data.run = run
 })
-
-const isHovered = ref(false)
-const customStyle = ref({})
-
-// Show/hide the resize handles when hovering
-const resizeHandleStyle = computed(() => ({
-visibility: isHovered.value ? 'visible' : 'hidden',
-width: '12px',
-height: '12px',
-}))
-
-// Font size control
-const currentFontSize = ref(14); // Default font size
-const minFontSize = 10;
-const maxFontSize = 24;
-const fontSizeStep = 2;
-
-const increaseFontSize = () => {
-  currentFontSize.value = Math.min(currentFontSize.value + fontSizeStep, maxFontSize);
-};
-
-const decreaseFontSize = () => {
-  currentFontSize.value = Math.max(currentFontSize.value - fontSizeStep, minFontSize);
-};
-
-// References to DOM elements
-const textContainer = ref(null)
-
-// Auto-scroll control
-const isAutoScrollEnabled = ref(true)
-
-// Access zoom functions from VueFlow
-const { zoomIn, zoomOut } = useVueFlow()
-
-// Disable zoom when interacting with the text container
-const disableZoom = () => {
-  zoomIn(0)
-  zoomOut(0)
-}
-
-// Enable zoom when not interacting
-const enableZoom = () => {
-  zoomIn(1)
-  zoomOut(1)
-}
-
-// Function to scroll to the bottom of the text container
-const scrollToBottom = () => {
-  nextTick(() => {
-    if (textContainer.value) {
-      textContainer.value.scrollTop = textContainer.value.scrollHeight
-    }
-  })
-}
-
-// Handle scroll events to toggle auto-scroll
-const handleScroll = () => {
-  if (textContainer.value) {
-    const { scrollTop, scrollHeight, clientHeight } = textContainer.value
-    if (scrollTop + clientHeight < scrollHeight) {
-      isAutoScrollEnabled.value = false
-    } else {
-      isAutoScrollEnabled.value = true
-    }
-  }
-}
-
-// Handle resize events
-const onResize = (event) => {
-  customStyle.value.width = `${event.width}px`
-  customStyle.value.height = `${event.height}px`
-}
 
 // Watch for changes and emit them upward
 watch(
   () => props.data,
   (newData) => {
     emit('update:data', { id: props.id, data: newData })
-    if (isAutoScrollEnabled.value) {
-      scrollToBottom()
-    }
   },
   { deep: true }
 )
-
-// Define five pastel colors suitable for a sticky note background.
-const pastelColors = ['#FFB3BA', '#FFDFBA', '#FFFFBA', '#BAFFC9', '#BAE1FF']
-const currentColorIndex = ref(0)
-const currentColor = computed(() => pastelColors[currentColorIndex.value])
-
-// Add computed properties for scroll bar colors based on current background color
-const scrollbarTrackColor = computed(() => currentColor.value)
-const scrollbarBorderColor = computed(() => currentColor.value)
-
-const cycleColor = () => {
-  currentColorIndex.value = (currentColorIndex.value + 1) % pastelColors.length
-  // Add this line to update the data style with the new color
-  props.data.style.backgroundColor = currentColor.value
-}
 </script>
-
 <style scoped>
 .node-container {
   background-color: var(--node-bg-color, #f7f3d7) !important;
@@ -221,7 +134,6 @@ const cycleColor = () => {
   color: var(--node-text-color);
   font-family: 'Roboto', sans-serif;
 }
-
 .note-node {
   /* Remove the hardcoded color override */
   /* --node-bg-color: #e8c547 !important; */
@@ -232,7 +144,6 @@ const cycleColor = () => {
   flex-direction: column;
   font-family: 'Arial', sans-serif;
 }
-
 .node-label {
   font-family: 'Roboto', sans-serif;
   font-size: 14px;
@@ -243,7 +154,6 @@ const cycleColor = () => {
   pointer-events: none;
   box-sizing: border-box;
 }
-
 .text-container {
   flex-grow: 1;
   padding: 2px;
@@ -254,7 +164,6 @@ const cycleColor = () => {
   box-sizing: border-box;
   overflow-y: hidden;
 }
-
 .note-textarea {
   width: 100%;
   height: 100%;
@@ -278,33 +187,27 @@ const cycleColor = () => {
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
-
 .note-textarea::placeholder {
   color: rgba(0,0,0,0.35);
   font-style: italic;
 }
-
 .note-textarea::-webkit-scrollbar {
   width: 4px; /* Even thinner scrollbar for more circular appearance */
 }
-
 .note-textarea::-webkit-scrollbar-track {
   background: transparent; /* Keep the background transparent */
   border-radius: 12px;
   margin: 8px 0; /* Add some margin to create space between thumb and edges */
 }
-
 .note-textarea::-webkit-scrollbar-thumb {
   background-color: rgba(82, 82, 91, 0.4); /* Neutral dark slate with transparency */
   border-radius: 10px; /* Rounded corners for circle effect */
   border: none; /* Remove border */
   min-height: 30px; /* Ensure minimum size for better visibility */
 }
-
 .note-textarea::-webkit-scrollbar-thumb:hover {
   background-color: rgba(82, 82, 91, 0.6); /* More opaque on hover */
 }
-
 /* Styles for the color cycle button */
 .color-cycle-btn {
   position: absolute;
@@ -323,12 +226,10 @@ const cycleColor = () => {
   font-size: 16px;
   outline: none;
 }
-
 .color-cycle-btn:focus {
   outline: none;
   box-shadow: none;
 }
-
 /* Font size control styles */
 .font-size-controls {
   position: absolute;
@@ -337,7 +238,6 @@ const cycleColor = () => {
   display: flex;
   gap: 4px;
 }
-
 .font-size-controls button {
   width: 24px;
   height: 24px;
@@ -352,7 +252,6 @@ const cycleColor = () => {
   font-size: 18px;
   color: #333;
 }
-
 .font-size-controls button:hover,
 .font-size-controls button:focus,
 .font-size-controls button:active {
